Moved reducers out of index files
