+++ src/reducers/selectedNavReducer/index.ts
@@ -1,16 +1,2 @@
-import { IAction } from "../../actions";
-import { IPutSelected } from "../../actions/selectedNav";
-import { PUT_SELECTED } from "../../actionTypes";
-
-export default function selectedNavReducer(
-  state: string[] = [],
-  action: IAction
-) {
-  switch (action.type) {
-    case PUT_SELECTED:
-      const { order } = action as IPutSelected;
-      return order;
-    default:
-      return state;
-  }
-}
+import selectedNavReducer from "./selectedNavReducer";
+export default selectedNavReducer;
