+++ src/reducers/userReducer/index.ts
@@ -1,19 +1,2 @@
-import { IAction, IActionUser } from "../../actions";
-import { INewUserObject } from "../../actions/authentication";
-import { NEW_USER_OBJECT, UPDATE_USER } from "../../actionTypes";
-import { IStateUsers } from "../index";
-
-export default function userReducer(state: IStateUsers = {}, action: IAction) {
-  switch (action.type) {
-    case NEW_USER_OBJECT:
-      return (action as INewUserObject).users;
-    case UPDATE_USER:
-      const users = { ...state };
-      users[
-        (action as IActionUser).user.userIdentifier
-      ] = (action as IActionUser).user;
-      return users;
-    default:
-      return state;
-  }
-}
+import userReducer from "./userReducer";
+export default userReducer;
